refactor: 패지키 구조 변경(dto, mapper)
