+++ src/main/java/roomescape/controller/dto/request/TimeRequest.java
@@ -1,13 +1,12 @@
-package roomescape.dto.response;
+package roomescape.controller.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
 
-public record TimeResponse(
-
-        long id,
-
+public record TimeRequest(
         @JsonFormat(pattern = "HH:mm")
+        @NotNull
         LocalTime startAt
 ) {
 }
